+++ src/snmp/smeerrtable/smeErrTable_data_get.cpp
@@ -206,7 +206,7 @@ smeErrTable_indexes_set(smeErrTable_rowreq_ctx *rowreq_ctx, long smeErrIndex_val
     /*
      * convert mib index to oid index
      */
-    rowreq_ctx->oid_idx.len = sizeof(rowreq_ctx->oid_tmp) / sizeof(oid);
+    rowreq_ctx->oid_idx.len = (int)(sizeof(rowreq_ctx->oid_tmp) / sizeof(oid));
     if(0 != smeErrTable_index_to_oid(&rowreq_ctx->oid_idx,
                                     &rowreq_ctx->tbl_idx)) {
         return MFD_ERROR;
@@ -257,7 +257,7 @@ SME system Id.
 int
 smeErrSystemId_map(char **mib_smeErrSystemId_val_ptr_ptr, size_t *mib_smeErrSystemId_val_ptr_len_ptr, char *raw_smeErrSystemId_val_ptr, size_t raw_smeErrSystemId_val_ptr_len, int allow_realloc)
 {
-    int converted_len;
+    size_t converted_len;
 
     netsnmp_assert(NULL != raw_smeErrSystemId_val_ptr);
     netsnmp_assert((NULL != mib_smeErrSystemId_val_ptr_ptr) && (NULL != mib_smeErrSystemId_val_ptr_len_ptr));
